+++ CHANGELOG.md
@@ -1,6 +1,12 @@
+## 0.1.7-dev.1
+
+-   Rough in all Filter constructors (isEq, isGt, etc).
+-   All `Filter` constructors accepts `dynamic` values, will add `Type` support in future.
+-   Transition to node-mock-server for the example.
+
 ## 0.1.6-dev.1
 
--   Complete Query
+-   Made jsonObjects available in `Response` (jinja renderMap requirements).
 
 ## 0.1.5-dev.1